+++ src/main/java/view/MainMenuFrame.java
@@ -1,3 +1,5 @@
+package view;
+
 import javax.swing.*;
 import java.awt.*;
